Move HTML printers to a new module
