+++ bokehjs/src/lib/models/tools/actions/examine_tool.ts
@@ -3,7 +3,8 @@ import type * as p from "core/properties"
 import * as icons from "styles/icons.css"
 import type {DialogView} from "../../ui/dialog"
 import {Dialog} from "../../ui/dialog"
-import {Examiner, HTMLPrinter} from "../../ui/examiner"
+import {Examiner} from "../../ui/examiner"
+import {ValuePrinter} from "models/ui/printers"
 import {HTML} from "../../dom/html"
 import type {View} from "core/build_views"
 import {build_view} from "core/build_views"
@@ -26,7 +27,7 @@ export class ExamineToolView extends ActionToolView {
     await super.lazy_initialize()
 
     const target = this.parent.model
-    const printer = new HTMLPrinter()
+    const printer = new ValuePrinter()
 
     const title_el = div()
     render(printer.to_html(target), title_el)
